requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
requests
openai
faiss-cpu
langchain
pypdf
tiktoken